CI: build with VS2019 toolset for Windows
This (compared to using the latest VS2022, as we currently do) increases
the chance that pymmcore will work with whatever version of the C++
runtime that gets loaded (possibly by other modules imported before us).
It may also help with installed by pip into a Conda environment, though
that is not a recommended use case.